improve(Finalizer): CCTP V1 finalizers can be renamed to be SVM specific
No EVM chains need CCTP V1 finalizers anymore so the top level functions can be renamed to reference SVM specifically
