+++ src/finalizer/utils/cctp/index.ts
@@ -1,4 +1,3 @@
-export * from "./l1ToL2";
-export * from "./l2ToL1";
+export * from "./l1ToSvmL2";
+export * from "./Svml2ToL1";
 export * from "./anyToAny";
-export * from "./svm";
